handle error + verify request with joi

+++ services/upload.js
@@ -1,25 +1,33 @@
-const fs = require('fs');
-const { addFileUploadRepo, getFileUploadRepo, deleteFileUploadRepo } = require("../repositories/upload")
+const fs = require("fs");
+const {
+  addFileUploadRepo,
+  getFileUploadRepo,
+  deleteFileUploadRepo,
+} = require("../repositories/upload");
 
 const addFileUploadService = (data) => {
-    return addFileUploadRepo(data)
-}
+  return addFileUploadRepo(data);
+};
 
 const getFileUploadService = (id) => {
-    return getFileUploadRepo(id);
-}
+  return getFileUploadRepo(id);
+};
 
 const deleteFileUploadService = async (id) => {
-    const meta = await getFileUploadRepo(id);
-    console.log(meta)
-    const path = `${process.cwd()}/uploads/${meta.filename}`
-    if(fs.existsSync(path)) {
-        fs.unlinkSync(path);
+  const meta = await getFileUploadRepo(id);
+  console.log(meta);
+  if (meta) {
+    const path = `${process.cwd()}/uploads/${meta.filename}`;
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
     }
-   
-    return deleteFileUploadRepo(id);
-}
+  }
+
+  return deleteFileUploadRepo(id);
+};
 
 module.exports = {
-    addFileUploadService, getFileUploadService, deleteFileUploadService
-}
+  addFileUploadService,
+  getFileUploadService,
+  deleteFileUploadService,
+};
